requirements.txt
SpeechRecognition==3.8.1
tk==0.1.0
textblob==0.15.3
matplotlib==3.4.2
mysql-connector-python==8.0.33
pyaudio==0.2.14
tkcalendar==2.1.1
pandas==2.0.3
numpy==1.24.3
